Moved `CoreDataFeedStore` to production target, along with the xcdatamodel and `CDFeed` and `CDFeedImage` types
