+++ FeedStoreChallenge/CoreData/CDFeed.swift
@@ -1,0 +1,14 @@
+//
+//  CDFeed.swift
+//  FeedStoreChallenge
+//
+//  Created by Bogdan P on 31/01/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import CoreData
+
+internal class CDFeed: NSManagedObject {
+	@NSManaged var timestamp: Date
+	@NSManaged var feed: NSOrderedSet
+}
