+++ FeedStoreChallenge/CoreData/CDFeedImage.swift
@@ -1,0 +1,17 @@
+//
+//  CDFeedImage.swift
+//  FeedStoreChallenge
+//
+//  Created by Bogdan P on 31/01/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import CoreData
+
+internal class CDFeedImage: NSManagedObject {
+	@NSManaged var id: UUID
+	@NSManaged var imageDescription: String?
+	@NSManaged var imageLocation: String?
+	@NSManaged var url: URL
+	@NSManaged var feed: CDFeed
+}
