+++ Tests/CoreDataFeedStoreTests.swift
@@ -5,137 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-import CoreData
-
-class CDFeedImage: NSManagedObject {
-	@NSManaged var id: UUID
-	@NSManaged var imageDescription: String?
-	@NSManaged var imageLocation: String?
-	@NSManaged var url: URL
-	@NSManaged var feed: CDFeed
-}
-
-extension CDFeedImage {
-	func populate(from feedImage: LocalFeedImage) -> CDFeedImage {
-		id = feedImage.id
-		imageDescription = feedImage.description
-		imageLocation = feedImage.location
-		url = feedImage.url
-		return self
-	}
-}
-
-class CDFeed: NSManagedObject {
-	@NSManaged var timestamp: Date
-	@NSManaged var feed: NSOrderedSet
-}
-
-extension CDFeed {
-	func populate(from feed: [LocalFeedImage], timestamp: Date, in context: NSManagedObjectContext) -> CDFeed {
-		self.feed = NSOrderedSet(array: feed.map { CDFeedImage(context: context).populate(from: $0) })
-		self.timestamp = timestamp
-		return self
-	}
-}
-
-extension CDFeed {
-	var fetchRequest: NSFetchRequest<CDFeed> {
-		NSFetchRequest<CDFeed>(entityName: CDFeed.entity().name!)
-	}
-}
-
-extension LocalFeedImage {
-	init?(from cacheFeedImage: CDFeedImage?) {
-		guard let cacheFeedImage = cacheFeedImage else { return nil }
-		
-		self.init(
-			id: cacheFeedImage.id,
-			description: cacheFeedImage.imageDescription,
-			location: cacheFeedImage.imageLocation,
-			url: cacheFeedImage.url
-		)
-	}
-}
-
-class CoreDataFeedStore: FeedStore {
-	
-	private let persistentContainer: NSPersistentContainer
-	private let managedContext: NSManagedObjectContext
-	private let dataModelName = "FeedDataModel"
-	private let devNullURL = URL(fileURLWithPath: "/dev/null")
-	
-	init() {
-		let model = NSManagedObjectModel(name: dataModelName, in: Bundle(for: CoreDataFeedStore.self))
-		persistentContainer = NSPersistentContainer(dataModelName: dataModelName, model: model, storeURL: devNullURL)
-		managedContext = persistentContainer.newBackgroundContext()
-	}
-	
-	func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		let context = self.managedContext
-		
-		context.perform {
-			self.deleteCache()
-			
-			completion(nil)
-		}
-	}
-	
-	func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		let context = self.managedContext
-		
-		context.perform {
-			self.deleteCache()
-			
-			let _ = CDFeed(context: context).populate(from: feed, timestamp: timestamp, in: context)
-			
-			try! context.save()
-			
-			completion(nil)
-		}
-	}
-	
-	func retrieve(completion: @escaping RetrievalCompletion) {
-		let context = self.managedContext
-		
-		context.perform {
-			guard let cachedFeed = try! context.fetch(CDFeed.fetchRequest()).first as? CDFeed else {
-				completion(.empty)
-				return
-			}
-			
-			completion(
-				.found(
-					feed: cachedFeed.feed.compactMap({ LocalFeedImage(from: $0 as? CDFeedImage) }),
-					timestamp: cachedFeed.timestamp
-				)
-			)
-		}
-	}
-	
-	private func deleteCache() {
-		if let feedCache = try! managedContext.fetch(CDFeed.fetchRequest()).first as? CDFeed {
-			managedContext.delete(feedCache)
-			try! managedContext.save()
-		}
-	}
-}
-
-private extension NSPersistentContainer {
-	convenience init(dataModelName: String, model: NSManagedObjectModel, storeURL: URL) {
-		let description = NSPersistentStoreDescription(url: storeURL)
-		self.init(name: dataModelName, managedObjectModel: model)
-		persistentStoreDescriptions = [description]
-		loadPersistentStores { _, _ in }
-	}
-}
-
-private extension NSManagedObjectModel {
-	convenience init(name: String, in bundle: Bundle) {
-		let modelURL = bundle.url(forResource: name, withExtension: "momd")!
-		self.init(contentsOf: modelURL)!
-	}
-}
-
 class CoreDataFeedStoreTests: XCTestCase, FeedStoreSpecs {
 	
 	func test_retrieve_deliversEmptyOnEmptyCache() {
